XWIKI-17968: Recomputation of average rating is wrong
  * Ensure to perform a float division

(cherry picked from commit c86b42fe9e24cf8e9fb24cfea3d2f8ed548ea206)

+++ xwiki-platform-core/xwiki-platform-ratings/xwiki-platform-ratings-api/src/main/java/org/xwiki/ratings/internal/SolrRatingsManager.java
@@ -480,7 +480,7 @@ public class SolrRatingsManager implements RatingsManager
                 offsetIndex += AVERAGE_COMPUTATION_BATCH_SIZE;
             } while (!ratings.isEmpty());
 
-            float newAverage = sumOfVotes / numberOfVotes;
+            float newAverage = Float.valueOf(sumOfVotes) / numberOfVotes;
             return this.getAverageRatingManager().resetAverageRating(entityReference, newAverage, numberOfVotes);
         } else {
             throw new RatingsException(AVERAGE_RATING_NOT_ENABLED_ERROR_MESSAGE);
